Update lib/ffi-rzmq.rb
This code should still require 'ffi' on Rubinius. We should not have
built in the assumption that the FFI constant would be available without
a require. Doing so makes it difficult for us to change how that constant
is made available. We're working on supporting the ruby-ffi gem now and
hiding our native implementation.